add the rank interval

+++ src/include/alg/SpaceInnerProduct.hpp
@@ -3,8 +3,8 @@
 
 namespace ReverseMIPS {
 
-    float InnerProduct(const float *pVect1, const float *pVect2, const int dim) {
-        float res = 0;
+    double InnerProduct(const double *pVect1, const double *pVect2, const int dim) {
+        double res = 0;
         for (unsigned i = 0; i < dim; i++) {
             res += pVect1[i] * pVect2[i];
         }
